% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{autoplot.plm}
\alias{autoplot.plm}
\title{Autoplot plm fit}
\usage{
\method{autoplot}{plm}(
  x,
  type = "rating_curve",
  param = NULL,
  transformed = F,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class "plm"}

\item{type}{a character denoting what type of plot should be drawn. Defaults to "rating_curve". Possible types are
\itemize{
   \item{"rating_curve"}{ to plot the rating curve.}
   \item{"rating_curve_mean"}{ to plot the posterior mean of the rating curve.}
   \item{"f"}{ to plot the power-law exponent.}
   \item{"beta"}{ to plot the random effect in the power-law exponent.}
   \item{"sigma_eps"}{ to plot the standard deviation on the data level.}
   \item{"residuals"}{ to plot the log residuals.}
   \item{"trace"}{ to plot trace plots of parameters given in param.}
   \item{"histogram"}{ to plot histograms of parameters given in param.}
}}

\item{param}{a character vector with the parameters to plot. Defaults to NULL and is only used if type is "trace" or "histogram". Allowed values are the parameters given in the model summary of x as well as "hyperparameters" or "latent_parameters" for specific groups of parameters.}

\item{transformed}{a logical value indicating whether the quantity should be plotted on a transformed scale used during the Bayesian inference. Defaults to FALSE.}

\item{title}{a character denoting the title of the plot. Defaults to NULL, i.e. no title.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\value{
returns an object of class ggplot2
}
\description{
Uses ggplot2 to plot plm object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
plm.fit <- plm(f,V316_river)
autoplot(plm.fit)
}
}
\seealso{
\code{\link{plm}} for fitting the plm model,\code{\link{summary.plm}} for summaries of model parameters, \code{\link{predict.plm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm}} to help visualize the full posterior distributions.
}
