% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{B_splines}
\alias{B_splines}
\title{B-splines in a generalized rating curve}
\usage{
B_splines(ZZ)
}
\arguments{
\item{ZZ}{A numeric matrix of dimension 1xn where n is number of osbervations. The input is calculated as follows:
(h-min(h)) divided by last element of the resulting vector, where h is stage observations.}
}
\value{
The function returns a linear combination of scaled B-spline basis functions for every stage observation.
}
\description{
A function to calculate the B-splines in a rating curve. When calculating error variance of log discharge in a rating curve the data depends on stage. It is modeled as an exponential of a B-splines curve of order 4,
with 2 interior knots and 6 basis functions.
}
