% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdrc-package.R
\docType{package}
\name{bdrc-package}
\alias{bdrc}
\alias{bdrc-package}
\title{bdrc: Bayesian Discharge Rating Curves}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fits a discharge rating curve based on the power-law and the generalized power-law from data on paired stage and discharge measurements in a given river using a Bayesian hierarchical model as described in Hrafnkelsson et al. (2020) \href{https://arxiv.org/abs/2010.04769}{arXiv:2010.04769}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sor16.github.io/bdrc}
  \item Report bugs at \url{https://github.com/sor16/bdrc/issues}
}

}
\author{
\strong{Maintainer}: Rafael Daníel Vias \email{raffidv@gmail.com} (\href{https://orcid.org/0009-0007-2601-6800}{ORCID})

Authors:
\itemize{
  \item Birgir Hrafnkelsson \email{birgirhr@hi.is} (\href{https://orcid.org/0000-0003-1864-9652}{ORCID}) [copyright holder]
  \item Solvi Rognvaldsson \email{solviro@gmail.com} (\href{https://orcid.org/0000-0002-4376-3361}{ORCID})
  \item Axel Orn Jansson \email{axelorn94@gmail.com}
}

}
\keyword{package}
