% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdrc.R
\docType{package}
\name{bdrc}
\alias{bdrc}
\title{bdrc - Bayesian Discharge Rating Curves}
\description{
The package implements the following Bayesian hierarchical discharge rating curve models for paired measurements of stage and discharge in rivers described in Hrafnkelsson et al.:
}
\details{
\code{\link{plm0}} - Power-law model with constant variance

\code{\link{plm}} - Power-law model with variance that may vary with stage

\code{\link{gplm0}} - Generalized power-law model with constant variance

\code{\link{gplm}} - Generalized power-law model with variance that may vary with stage
}
