% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgplm0.R
\name{bgplm0}
\alias{bgplm0}
\title{Bayesian Generalized Power Law Model with Constant Variance}
\usage{
bgplm0(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  forcepoint = rep(FALSE, nrow(data)),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula", with discharge column name as response and stage column name as a covariate.The details of model specification are given under "Details".}

\item{data}{data.frame containing the variables specified in formula}

\item{c_param}{stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data}

\item{h_max}{maximum stage to which the rating curve should extrapolate for. If NULL, the maximum stage value in data is selected as an upper bound.}

\item{forcepoint}{A boolean vector of the same length as the number of rows in data. If an element at index i is TRUE it indicates that the rating curve should be forced through the i-th measurement. Use with care, as this will strongly influence the resulting rating curve.}
}
\value{

}
\description{
bgplm0 is used to fit a rating curve for paired measurements of stage and discharge using a Bayesian Power Law Model with constant variance as described in Hrafnkelsson et al.
}
\examples{
data(V316_river)
f <- Q~W
bgplm0.fit <- bgplm0(f,V316_river)
summary(bgplm0.fit)
plot(bgplm0.fit)
bgplm0.fit_known_c <- bgplm0(f,sim_dat,c_param=)
summary(bgplm0.fit)
plot(bgplm0.fit)
}
\references{
B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv
preprint 2010.04769
}
\seealso{
\code{\link{summary.bgplm0}} for summaries, \code{\link{predict.bgplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bgplm0.plot}} to help visualize the full posterior distributions.
}
