% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplm0.R
\name{bplm0}
\alias{bplm0}
\title{Fitting discharge rating curves using Bayesian Power Law Model with constant variance}
\usage{
bplm0(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  forcepoint = rep(FALSE, nrow(data)),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula", with discharge column name as response and stage column name as a covariate.The details of model specification are given under "Details".}

\item{data}{data.frame containing the variables specified in formula}

\item{c_param}{stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data}

\item{h_max}{maximum stage to which the rating curve should extrapolate for. If NULL, the maximum stage value in data is selected as an upper bound.}

\item{forcepoint}{A boolean vector of the same length as the number of rows in data. If an element at index i is TRUE it indicates that the rating curve should be forced through the i-th measurement. Use with care, as this will strongly influence the resulting rating curve.}
}
\value{
bplm0 returns an object of class "bplm0"\cr\cr
The function summary is used to obtain and print a summary of the model.\cr\cr
An object of class "bplm0" is a list containing the following components: \cr

\item{\code{rating_curve}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of the rating curve.}
\item{\code{rating_curve_mean}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of the mean of the rating curve.}
\item{\code{param_summary}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of latent- and hyperparameters.}
\item{\code{DIC_summary}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of the Deviance Information Criterion.}
\item{\code{rating_curve_posterior}}{a matrix containing the full thinned posterior samples of the posterior distribution of the rating curve (excluding burn-in).}
\item{\code{rating_curve_mean_posterior}}{a matrix containing the full thinned posterior samples of the posterior distribution of the mean of the rating curve (excluding burn-in).}
\item{\code{a_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{a}.}
\item{\code{b_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{b}.}
\item{\code{c_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{c}.}
\item{\code{sigma_eps_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{\sigma_{\epsilon}}.}
\item{\code{formula}}{object of type "formula" provided by the user.}
\item{\code{data}}{data provided by the user.}
\item{\code{run_info}}{Information about the specific parameters used in the MCMC chain.}
}
\description{
bplm0 is used to fit a rating curve for paired measurements of stage and discharge using a Bayesian Power Law Model with constant variance as described in Hrafnkelsson et al.
}
\examples{
data(V316_river)
f <- Q~h
bplm0.fit <- bplm0(f,V316_river)
summary(bplm0.fit)
plot(bplm0.fit)
bplm0.fit_known_c <- bplm0(f,sim_dat,c_param=)
summary(bplm0.fit)
plot(bplm0.fit)
}
\references{
B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv
preprint 2010.04769
}
\seealso{
\code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm0.plot}} to help visualize the full posterior distributions.
}
