% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{calc_waic}
\alias{calc_waic}
\title{calc_waic is a function that takes in a model object (m) and data (d) and returns a list containing the elements used to calculate WAIC.}
\usage{
calc_waic(m, d)
}
\description{
calc_waic is a function that takes in a model object (m) and data (d) and returns a list containing the elements used to calculate WAIC.
}
