% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{compare_all}
\alias{compare_all}
\title{Fit and Compare All Models in the bdrc Package}
\usage{
compare_all(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  parallel = TRUE,
  num_cores = NULL,
  forcepoint = rep(FALSE, nrow(data)),
  verbose = TRUE
)
}
\arguments{
\item{formula}{An object of class "formula", with discharge column name as response and stage column name as a covariate, i.e. of the form \code{y} ~ \code{x} where \code{y} is discharge in m\eqn{^3/}s and \code{x} is stage in m (the data must be in the correct units). To include discharge measurement-error data, use the form \code{y | y_error ~ x}, where \code{y_error} is the column name for discharge measurement errors.}

\item{data}{A data.frame containing the variables specified in formula.}

\item{c_param}{The largest stage value for which there is zero discharge. If NULL (default), it is treated as unknown in the model and inferred from the data.}

\item{h_max}{The maximum stage to which the rating curve should extrapolate to. If NULL (default), the maximum stage value in the data is selected as an upper bound.}

\item{parallel}{A logical value indicating whether to run the MCMC in parallel or not. Defaults to TRUE.}

\item{num_cores}{An integer between 1 and 4 (number of MCMC chains) indicating how many cores to use. Only used if parallel=TRUE. If NULL, the number of cores available on the device is detected automatically.}

\item{forcepoint}{A logical vector of the same length as the number of rows in data. If an element at index \eqn{i} is TRUE it indicates that the rating curve should be forced through the \eqn{i}-th measurement. Use with care, as this will strongly influence the resulting rating curve.}

\item{verbose}{A logical value indicating whether to print progress and diagnostic information. If `TRUE`, the function will print messages as it runs. If `FALSE`, the function will run silently. Default is `TRUE`.}
}
\value{
An object of class "compare" containing a data frame with comparison metrics for all fitted models. See ?compare for details.
}
\description{
This function fits all applicable models (plm0, plm, gplm0, gplm) to the given data,
including measurement error models if applicable, and then compares them using WAIC.
}
\examples{
\dontrun{
set.seed(1)
data(provo)
compare_result <- compare_all(Q | Q_sigma ~ W, provo, num_cores = 2)
plot(compare_results)
}
}
