% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Densevalm11.R
\name{Densevalm11}
\alias{Densevalm11}
\title{Density evaluation for model1}
\usage{
Densevalm11(th, RC)
}
\arguments{
\item{th}{A vector with length 9 containing parameters}

\item{RC}{A list containing prior parameters, matrices and the data.}
}
\value{
Returns a list containing predictive values of the parameters drawn out of the evaluated density.
}
\description{
Evaluates the log density of the posterior distribution of the parameters of model2BH.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}

