% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_draws.R
\name{gather_draws}
\alias{gather_draws}
\title{Gather MCMC chain draws to data.frame on a long format}
\usage{
gather_draws(mod, ..., transformed = F)
}
\arguments{
\item{mod}{an object of class "bplm0","bplm","bgplm0" or "bgplm".}

\item{...}{character vectors of parameters or other output from the model output of interest. Also accepts "latent_parameters" and "hyperparameters".}

\item{transformed}{boolean value determining whether the parameter is to be represented on the transformed scale used for sampling in the MCMC chain or the original scale. Defaults to FALSE.}
}
\value{
Data frame with columns
\code{chain}
\code{iter}
\code{param}
\code{value}
}
\description{
Useful to convert MCMC chain draws of particular parameters or output from the model object to a long format for further data wrangling
}
\examples{
data(sim_dat)
f <- Q~W
bplm0.fit <- bplm0(f,sim_dat)
summary(bplm9.fit)
plot(bplm0.fit,type='rating_curve')
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson, & Sigurdur M. Gardarsson. (2020). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling.
}
\seealso{
\code{\link{bplm0}},\code{\link{bplm}},\code{\link{bgplm0}},\code{\link{bgplm}} for further information on parameters
}
