% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report.gplm}
\alias{get_report.gplm}
\title{Get gplm fit report as pdf file}
\usage{
\method{get_report}{gplm}(x, path = NULL, type = 1, ...)
}
\arguments{
\item{x}{an object of class "gplm".}

\item{path}{a valid path to store the pdf report. Defaults to save report at working directory as report.pdf.}

\item{type}{an integer denoting what type of report is to be produced. Only type 1 is permissible for an object of class "gplm" which produces a two page report consisting of a plot collage, a parameter summary table and a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Save a pdf report of an object
}
\examples{
\dontrun{
data(bunnerviken)
f <- Q~W
gplm.fit <- gplm(f,bunnerviken)
get_report(gplm.fit)
}
}
\seealso{
\code{\link{gplm}} for fitting the gplm model,\code{\link{summary.gplm}} for summaries, \code{\link{predict.gplm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm}} to help visualize the full posterior distributions.
}
