% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report.gplm0}
\alias{get_report.gplm0}
\title{Get gplm0 fit report as pdf file}
\usage{
\method{get_report}{gplm0}(x, path = NULL, type = 1, ...)
}
\arguments{
\item{x}{an object of class "gplm0".}

\item{path}{a valid path to store the pdf report. Defaults to save report at working directory as report.pdf.}

\item{type}{an integer denoting what type of report is to be produced. Only type 1 is permissible for an object of class "gplm0" which produces a two page report consisting of a plot collage, a parameter summary table and a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Save a pdf report of an object
}
\examples{
\dontrun{
data(flotemarken)
f <- Q~W
gplm0.fit <- gplm0(f,flotemarken)
get_report(gplm0.fit)
}
}
\seealso{
\code{\link{gplm0}} for fitting the gplm0 model,\code{\link{summary.gplm0}} for summaries, \code{\link{predict.gplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm0}} to help visualize the full posterior distributions.
}
