% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report.plm}
\alias{get_report.plm}
\title{Get plm fit report as pdf file}
\usage{
\method{get_report}{plm}(x, path = NULL, type = 1, ...)
}
\arguments{
\item{x}{an object of class "plm".}

\item{path}{a valid path to store the pdf report. Defaults to save report at working directory as report.pdf.}

\item{type}{an integer denoting what type of report is to be produced. Only type 1 is permissible for an object of class "plm" which produces a two page report consisting of a plot collage, a parameter summary table and a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Save a pdf report of an object
}
\examples{
\dontrun{
data(lisjobacken)
f <- Q~W
plm.fit <- plm(f,lisjobacken)
get_report(plm.fit)
}
}
\seealso{
\code{\link{plm}} for fitting the plm model,\code{\link{summary.plm}} for summaries, \code{\link{predict.plm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm}} to help visualize the full posterior distributions.
}
