% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report.plm0}
\alias{get_report.plm0}
\title{Get plm0 fit report}
\usage{
\method{get_report}{plm0}(x, path = NULL, type = 1, ...)
}
\arguments{
\item{x}{an object of class "plm0".}

\item{path}{a valid path to store the pdf report. Defaults to save report at working directory as report.pdf.}

\item{type}{an integer denoting what type of report is to be produced. Only type 1 is permissible for an object of class "plm0" which produces a two page report consisting of a plot collage, a parameter summary table and a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Save a pdf report of an object
}
\examples{
\dontrun{
data(halla)
f <- Q~W
plm0.fit <- plm0(f,halla)
get_report(plm0.fit)
}
}
\seealso{
\code{\link{plm0}} for fitting the plm0 model,\code{\link{summary.plm0}} for summaries, \code{\link{predict.plm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm0}} to help visualize the full posterior distributions.
}
