% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report}
\alias{get_report}
\title{Get report}
\usage{
get_report(x, path = NULL, type = 1, ...)
}
\arguments{
\item{x}{an object of class  "tournament", "plm0", "plm", "gplm0" or "gplm".}

\item{path}{a valid path to store the pdf report. Defaults to save report at working directory as report.pdf.}

\item{type}{an integer denoting what type of report is to be produced. Defaults to type 1. Only type 1 is permissible for an object of class "plm0", "plm", "gplm0" or "gplm". Possible types are
\itemize{
   \item{1}{ produces a two page report consisting of a plot collage, a parameter summary table and a tabular rating curve, from the results of a model object, or if input class is "tournament", to the most approriate model, as determined by \code{\link{tournament}}.}
   \item{2}{ produces a 10 page report from the results of all four models; \code{\link{plm0}}, \code{\link{plm}}, \code{\link{gplm0}} and \code{\link{gplm}}. The first four pages show a plot collage and the parameter summary table for the four models. The fifth page is a model comparison page, the sixth is a covergence diagnostics page and the last four pages show the histograms of the estimated parameters in the four models.}
}}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Save a pdf report of an object
}
\examples{
\dontrun{
data(bunnerviken)
f <- Q~W
t_obj <- tournament(f,bunnerviken)
get_report(t_obj)
}
}
\seealso{
\code{\link{tournament}} for running a tournament,\code{\link{summary.tournament}} for summaries.
}
