% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report_pages.gplm0}
\alias{get_report_pages.gplm0}
\title{Get gplm0 fit report pages}
\usage{
\method{get_report_pages}{gplm0}(x, type = 1, ...)
}
\arguments{
\item{x}{an object of class "gplm0".}

\item{type}{an integer to determine whether to get the pages from report type 1 or 2. Only type 1 is permissible for an object of class "gplm0" which produces a list of two grob objects, the first grob object consists of a plot collage and parameter summary table, the second a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Get a list of the grob objects used by \code{\link{get_report}} to create a pdf report.
}
\examples{
\dontrun{
data(flotemarken)
f <- Q~W
gplm0.fit <- gplm0(f,flotemarken)
grob_object_list <- get_report_pages(gplm0.fit)
}
}
\seealso{
\code{\link{gplm0}} for fitting the gplm0 model,\code{\link{summary.gplm0}} for summaries, \code{\link{predict.gplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm0}} to help visualize the full posterior distributions.
}
