% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report_pages.plm}
\alias{get_report_pages.plm}
\title{Get report pages}
\usage{
\method{get_report_pages}{plm}(x, type = 1, ...)
}
\arguments{
\item{x}{an object of class "plm".}

\item{type}{an integer denoting what type of report is to be produced. Only type 1 is permissible for an object of class "plm". Type 1 report consists of two pages. The first page contains a panel of four plots and a summary of the posterior distributions of the paraters. The second one contains a tabular prediction of discharge on an equally spaced grid of stages.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Get a list of the grob objects the make up the pages of the report generated with \code{\link{get_report}}.
}
\examples{
\dontrun{
data(lisjobacken)
f <- Q~W
plm.fit <- plm(f,lisjobacken)
grob_object_list <- get_report_pages(plm.fit)
}
}
\seealso{
\code{\link{get_report}} for generating and saving a report.
}
