% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report_pages.tournament}
\alias{get_report_pages.tournament}
\title{Get tournament report pages}
\usage{
\method{get_report_pages}{tournament}(x, type = 1, ...)
}
\arguments{
\item{x}{an object of class "tournament".}

\item{type}{an integer denoting what type of report is to be produced. Defaults to type 1. Possible types are
\itemize{
   \item{1}{ produces a list of two grob objects, the first grob object consists of a plot collage and parameter summary table, the second a tabular rating curve from the results of the most approriate model for a given data set, as determined by \code{\link{tournament}}.}
   \item{2}{ produces a list of ten grob objects, the first four grob objects consist of plot collage and parameter summary table for the four models, the fifth object consists of model comparison plots and tables, the sixth consists of convergence diagnostics plots, and the final four grob objects consist of histogram plots of the estmated parameters for the four models.}
}}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Get a list of the grob objects used by \code{\link{get_report}} to create a pdf report.
}
\examples{
\dontrun{
data(bunnerviken)
f <- Q~W
t_obj <- tournament(f,bunnerviken)
grob_object_list <- get_report_pages(t_obj)
}
}
\seealso{
\code{\link{tournament}} for running a tournament,\code{\link{summary.tournament}} for summaries.
}
