% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{h_unobserved}
\alias{h_unobserved}
\title{Unobserved stages}
\usage{
h_unobserved(RC, h_min = NA, h_max = NA)
}
\arguments{
\item{RC}{list which must include the stage values h,h_min and h_max}

\item{h_min}{minimum stage of rating curve.}

\item{h_max}{maximum stage of rating curve.}
}
\value{
h_unobserved returns a list of vectors, h_u and h_u_tild. h_u is a vector of unobserved stage values
needed to make an equally spaced grid of stages. h_u_tild is a vector which is calculated by h_u-min(h_unique) needed to input into B_splines.
The unobserved stages are lower or higher than that of the data, take the same value in h_u_tild as the minimum value and maximum value of the
data respectively. This is done to ensure constant variance below and above observed data.
}
\description{
h_unobserved returns the stages that are needed to make an equally spaced grid of stages from data of stages.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
