% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halla}
\alias{halla}
\title{Halla river in Sweden}
\format{
A data frame with columns:
\describe{
 \item{W}{Measurements of river stage}
 \item{Q}{Measurements of river discharge}
}
}
\source{
Swedish Meteorological and Hydrological Institute \url{http://www.smhi.se}
}
\usage{
halla
}
\description{
Paired stage (W) and discharge measurements (Q) from Halla river in Sweden
}
\keyword{datasets}
