% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jokdal}
\alias{jokdal}
\title{Jokulsa a Dal gauging station in Iceland}
\format{
A data frame with columns:
\describe{
 \item{\code{W}}{Measurements of water stage in meters}
 \item{\code{Q}}{Measurements of water discharge in cubic meters per second}
}
}
\source{
Icelandic Meteorological Office, Landsvirkjun - the National Power Company of Iceland, and the Icelandic Road and Coastal Administration.
}
\usage{
jokdal
}
\description{
Data on discharge and stage from Jokulsa a Dal gauging station in Iceland
}
\keyword{datasets}
