% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{minimal.plm0}
\alias{minimal.plm0}
\alias{minimal.plm}
\alias{minimal.gplm0}
\alias{minimal.gplm}
\title{Extract a reduced object without posterior samples}
\usage{
\method{minimal}{plm0}(x, ...)

\method{minimal}{plm}(x, ...)

\method{minimal}{gplm0}(x, ...)

\method{minimal}{gplm}(x, ...)
}
\arguments{
\item{x}{an object of class "plm0", "plm", "gplm0" or "gplm".}

\item{...}{not used in this function}
}
\description{
Obtain an object with posterior summaries only to reduce object size
}
\section{Functions}{
\itemize{
\item \code{minimal(plm0)}: Reduced plm0 object

\item \code{minimal(plm)}: Reduced plm object

\item \code{minimal(gplm0)}: Reduced gplm0 object

\item \code{minimal(gplm)}: Reduced gplm object

}}
\seealso{
\code{\link{plm0}}, \code{\link{plm}}, \code{\link{gplm0}}, \code{\link{gplm}} for fitting a discharge rating curve and \code{\link{summary.plm0}}, \code{\link{summary.plm}}, \code{\link{summary.gplm0}} and \code{\link{summary.gplm}} for summaries. It is also useful to look at \code{\link{plot.plm0}}, \code{\link{plot.plm}}, \code{\link{plot.gplm0}} and \code{\link{plot.gplm}} to help visualize all aspects of the fitted discharge rating curve.
}
