% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{minimal.tournament}
\alias{minimal.tournament}
\title{Reduce tournament size by stripping plm objects of posterior samples}
\usage{
\method{minimal}{tournament}(x, ...)
}
\arguments{
\item{x}{an object of class "tournament"}

\item{...}{not used in this function}
}
\description{
Obtain a tournament whose model objects include with posterior summaries only to reduce object size
}
\seealso{
\code{\link{tournament}} to run a discharge rating curve tournament, \code{\link{summary.tournament}} for summaries and \code{\link{plot.tournament}} for tournament visualization.
}
