% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model1BH.R
\name{model1BH}
\alias{model1BH}
\title{Rating curve model1}
\usage{
model1BH(clean, country = "Iceland", Wmin = "", Wmax = "")
}
\arguments{
\item{clean}{List that is the output of \code{\link{clean}} function.}

\item{country}{A string of t the country the prior parameters should be defined, default value is Iceland.}

\item{Wmin}{positive numeric value for the lowest stage the user wants to calculate a rating curve. If input is an empty string (default) Wmin will
automatically be set to c_hat.}

\item{Wmax}{positive numeric value for the highest stage the user wants to calculate a rating curve. If input is an empty string (default) Wmax will
automatically be set to the maximum stage of the data.}
}
\value{
List containing information on the calculated rating curve, parameters varappr and c_hat, the matrix mu and
the data frames observedData, completePrediction, observedPrediction , TableOfData, FitTable, LowerTable, UpperTable, plotTable.
}
\description{
Takes in the cleaned data from the \code{\link{clean}} function
and calculates a rating curve using model 1.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
\seealso{
\code{\link{clean}}
}

