% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model2BH.R
\name{model2BH}
\alias{model2BH}
\title{Calculates the rating curve using model2}
\usage{
model2BH(clean, country = "Iceland", Wmin = "", Wmax = "")
}
\arguments{
\item{clean}{List that is the output of \code{\link{clean}} function.}

\item{country}{Name of the country the prior parameters should be defined for, default value is "Iceland".}

\item{Wmin}{Positive numeric value for the lowest stage the user wants to calculate a rating curve. If input is an empty string (default) Wmin will
automatically be set to c_hat.}

\item{Wmax}{Positive numeric value for the highest stage the user wants to calculate a rating curve. If input is an empty string (default) Wmax will
automatically be set to the maximum stage of the data.}
}
\value{
List containing information on the calculated rating curve,
the data frames observedData, betaData, completePrediction, observedPrediction, TableOfData, FitTable, LowerTable, UpperTable, plotTable.
}
\description{
This function takes in clean data from the \code{\link{clean}} function
and calculates the rating curve using model 2.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
\seealso{
\code{\link{clean}}
}

