% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norn}
\alias{norn}
\title{Norn gauging station in Sweden}
\format{
A data frame with columns:
\describe{
 \item{\code{W}}{Measurements of water stage in meters}
 \item{\code{Q}}{Measurements of water discharge in cubic meters per second}
}
}
\source{
Swedish Meteorological and Hydrological Institute.
}
\usage{
norn
}
\description{
Data on discharge and stage from Norn gauging station in Sweden.
}
\keyword{datasets}
