% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm0.R
\name{plm0}
\alias{plm0}
\title{Power-law model with constant log-error variance.}
\usage{
plm0(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  parallel = TRUE,
  num_cores = NULL,
  forcepoint = rep(FALSE, nrow(data)),
  verbose = TRUE
)
}
\arguments{
\item{formula}{An object of class "formula", with discharge column name as response and stage column name as a covariate, i.e. of the form \code{y} ~ \code{x} where \code{y} is discharge in m\eqn{^3/}s and \code{x} is stage in m (the data must be in the correct units). To include discharge measurement-error data, use the form \code{y | y_error ~ x}, where \code{y_error} is the column name for discharge measurement errors.}

\item{data}{A data.frame containing the variables specified in formula.}

\item{c_param}{The largest stage value for which there is zero discharge. If NULL (default), it is treated as unknown in the model and inferred from the data.}

\item{h_max}{The maximum stage to which the rating curve should extrapolate to. If NULL (default), the maximum stage value in the data is selected as an upper bound.}

\item{parallel}{A logical value indicating whether to run the MCMC in parallel or not. Defaults to TRUE.}

\item{num_cores}{An integer between 1 and 4 (number of MCMC chains) indicating how many cores to use. Only used if parallel=TRUE. If NULL, the number of cores available on the device is detected automatically.}

\item{forcepoint}{A logical vector of the same length as the number of rows in data. If an element at index \eqn{i} is TRUE it indicates that the rating curve should be forced through the \eqn{i}-th measurement. Use with care, as this will strongly influence the resulting rating curve.}

\item{verbose}{A logical value indicating whether to print progress and diagnostic information. If `TRUE`, the function will print messages as it runs. If `FALSE`, the function will run silently. Default is `TRUE`.}
}
\value{
plm0 returns an object of class "plm0". An object of class "plm0" is a list containing the following components:
\describe{
  \item{\code{posterior}}{A list containing the full posterior samples for various parameters and derived quantities.}
  \item{\code{summary}}{A list containing summary statistics (2.5\%, 50\%, 97.5\% percentiles) for various parameters and derived quantities.}
  \item{\code{log_likelihood}}{A list containing log-likelihood related quantities including WAIC and DIC.}
  \item{\code{diagnostics}}{A list containing MCMC diagnostics including autocorrelation and acceptance rate.}
  \item{\code{formula}}{The formula provided by the user.}
  \item{\code{data}}{The data provided by the user, ordered by stage.}
  \item{\code{run_info}}{Information about the input arguments and specific parameters used in the MCMC chain.}
}
When discharge measurement-error data is included, the posterior and summary components also include Q_true (the estimated true discharge observations).
}
\description{
plm0 is used to fit a discharge rating curve for paired measurements of stage and discharge using a power-law model with constant log-error variance as described in Hrafnkelsson et al. (2022). It also supports discharge measurement-error data. See "Details" for a more in depth description of the model.
}
\details{
The power-law model, which is commonly used in hydraulic practice, is of the form
\deqn{Q=a(h-c)^{b}}
where \eqn{Q} is discharge, \eqn{h} is stage and \eqn{a}, \eqn{b} and \eqn{c} are unknown constants.\cr\cr
The power-law model is here inferred by using a Bayesian hierarchical model. The model is on a logarithmic scale
\deqn{\log(Q_i) = \mu_i + \varepsilon_i\hspace{20mm}}
\deqn{\hspace{10mm}\mu_i = \log(a) + b \log(h_i - c)}
for \eqn{i = 1,...,n}, where \eqn{\varepsilon_i} follows a normal distribution with mean zero and constant variance \eqn{\sigma_\varepsilon^2}.\cr\cr
When measurement-uncertainty data is included, the model accounts for the uncertainty in the discharge measurements, and \eqn{\sigma_\varepsilon^2} captures the remaining structural uncertainty. The measurement-uncertainty datum \eqn{Q_{\text{SE},i}} corresponding to an observed discharge \eqn{Q_{\text{OBS},i}} is assumed to be the standard deviation of a lognormal distribution with median value at the true observations, \eqn{Q_{\text{TRUE},i}}. Then the model can be summarized as
\deqn{\log(Q_{\text{OBS},i})\sim \mathcal{N}(\log(Q_{\text{TRUE},i}),\tau^2_i),}
\deqn{\log(Q_{\text{TRUE},i})\sim \mathcal{N}(\mu_i,\sigma_\varepsilon^2),\hspace{17mm}}
where \eqn{\tau^2_i} is the variance of the Gaussian measurement-error density. An estimate for \eqn{\tau_i^2} is computed with \eqn{\hat{\tau}^2_i=\log(1+(Q_{\text{SE},i}/\mu_{Q,i})^2)}, where \eqn{\mu_{Q,i}} is the mean of the lognormal density with median value \eqn{Q_{\text{OBS},i}} and standard deviation \eqn{Q_{\text{SE},i}}, and is computed as
\deqn{\mu_{Q,i}=\sqrt{\frac{Q_{\text{OBS},i}^2 + Q_{\text{OBS},i}\sqrt{Q_{\text{OBS},i}^2+4Q_{\text{SE},i}^2}}{2}}.}
Both numerical and categorical measurement-uncertainty data are supported:\cr
\itemize{
  \item Numerical data: Standard deviations of the lognormal measurment-error densities.
  \item Categorical data: USGS discharge measurement quality codes, which are automatically transformed into numerical values as follows:
    \itemize{
      \item E (Excellent): Within 2\% of the actual flow
      \item G (Good): Within 5\% of the actual flow
      \item F (Fair): Within 8\% of the actual flow
      \item P (Poor): Greater than 8\% of the actual flow
    }
}
An efficient posterior simulation is achieved by sampling from the joint posterior density of the hyperparameters of the model, and then sampling from the density of the latent parameters conditional on the hyperparameters (Hrafnkelsson et al., 2022).\cr\cr
}
\examples{
\donttest{
data(skogsliden)
data(provo)
data(mokelumne)
set.seed(1)

# Without measurement error data
plm0.fit <- plm0(formula = Q ~ W, data = skogsliden, num_cores = 2)
summary(plm0.fit)

# With numerical measurement error data
plm0_me.fit <- plm0(formula = Q | Q_sigma ~ W, data = provo, num_cores = 2)
summary(plm0_me.fit)

# With categorical measurement error data
plm0_me_cat.fit <- plm0(formula = Q | Q_quality ~ W, data = mokelumne, num_cores = 2)
summary(plm0_me_cat.fit)
}
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., and Rubin, D. B. (2013). Bayesian Data Analysis, Third Edition. Chapman & Hall/CRC Texts in Statistical Science. Taylor & Francis. doi: https://doi.org/10.1201/b16018

Hrafnkelsson, B., Sigurdarson, H., Rögnvaldsson, S., Jansson, A. Ö., Vias, R. D., and Gardarsson, S. M. (2022). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling, Environmetrics, 33(2):e2711. doi: https://doi.org/10.1002/env.2711

Spiegelhalter, D., Best, N., Carlin, B., Van Der Linde, A. (2002). Bayesian measures of model complexity and fit. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 64(4), 583–639. doi: https://doi.org/10.1111/1467-9868.00353

Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and widely applicable information criterion in singular learning theory. J. Mach. Learn. Res. 11, 3571–3594.
}
\seealso{
\code{\link{summary.plm0}}, \code{\link{predict.plm0}}, \code{\link{plot.plm0}}, \code{\link{spread_draws}}, \code{\link{gather_draws}}.
}
