% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{plot.bgplm0}
\alias{plot.bgplm0}
\title{Plot bplm0 fit}
\usage{
\method{plot}{bgplm0}(x, type = "rating_curve", ..., transformed = F)
}
\arguments{
\item{x}{an object of class "bplm0"}

\item{type}{a character to denote the plot type. One of 'trace','histogram','rating_curve','rating_curve_mean','residuals'}
}
\value{
bplm0 returns an object of class "bplm0"\cr\cr
}
\description{
Print the results of a  object
}
\examples{
data(V316_river)
f <- Q~W
bplm0.fit <- bplm0(f,V316_river)
plot(bplm0.fit)
}
\seealso{
\code{\link{bplm0}} for fitting the bplm0 model,\code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm0.plot}} to help visualize the full posterior distributions.
}
