% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_methods.R
\name{plot.compare}
\alias{plot.compare}
\title{Plot Method for compare Objects}
\usage{
\method{plot}{compare}(x, ...)
}
\arguments{
\item{x}{An object of class "compare", typically created by the compare() function.}

\item{...}{Additional arguments passed to plot (not used).}
}
\value{
A ggplot2 object representing the forest plot with the following elements:
  \itemize{
    \item White circles showing the WAIC estimate for each model.
    \item Black dots showing the within-sample prediction error ( WAIC - 2 * pWAIC ).
    \item Black error bars showing ±1 standard error of the WAIC estimate (SE).
    \item Gray error bars and triangles showing ±1 standard error (dSE) of the WAIC difference (dWAIC) between the best model and the model directly below.
    \item A vertical dashed line highlighting the WAIC of the best model.
  }
}
\description{
This function creates a forest plot of model comparison results based on WAIC (Widely Applicable Information Criterion).
}
\details{
The plot visualizes model comparison based on WAIC. Models are ordered from best (lowest WAIC)
at the top to worst at the bottom. The x-axis represents the deviance scale, where lower values
indicate better predictive accuracy.

The gray error bars around the triangles allow for quick visual of the uncertainty in the estimated differences in WAIC between each model and the model with the lowest WAIC.

See ?compare for details on WAIC calculation and interpretation.
}
\examples{
\dontrun{
set.seed(1)
data(krokfors)
plm0.fit <- plm0(Q ~ W, krokfors, num_cores = 2)
gplm.fit <- gplm(Q ~ W, krokfors, num_cores = 2)
compare_results <- compare(plm0.fit, gplm.fit)
plot(comparison_result)
}
}
