% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{plot.gplm0}
\alias{plot.gplm0}
\title{Plot gplm0 fit}
\usage{
\method{plot}{gplm0}(x, type = "rating_curve", param = NULL, transformed = F, ...)
}
\arguments{
\item{x}{an object of class "gplm0"}

\item{type}{a character denoting what type of plot should be drawn. Defaults to "rating_curve". Possible types are
\itemize{
   \item{"rating_curve"}{ to plot the rating curve.}
   \item{"rating_curve_mean"}{ to plot the posterior mean of the rating curve.}
   \item{"f"}{ to plot the power-law exponent.}
   \item{"beta"}{ to plot the random effect in the power-law exponent.}
   \item{"sigma_eps"}{ to plot the standard deviation on the data level.}
   \item{"residuals"}{ to plot the log residuals.}
   \item{"trace"}{ to plot trace plots of parameters given in param.}
   \item{"histogram"}{ to plot histograms of parameters given in param.}
}}

\item{param}{a character vector with the parameters to plot. Defaults to NULL and is only used if type is "trace" or "histogram". Allowed values are the parameters given in the model summary of x as well as "hyperparameters" or "latent_parameters" for specific groups of parameters.}

\item{transformed}{a logical value indicating whether the quantity should be plotted on a transformed scale used during the Bayesian inference. Defaults to FALSE.}

\item{...}{further arguments passed to other methods. Currently supports:
\itemize{
  \item{"title"}{ a character denoting the title of the plot}
  \item{"xlim"}{ numeric vector of length 2, denoting the limits on the x axis of the plot. Only active for types "rating_curve","rating_curve_mean","f","beta","sigma_eps","residuals".}
  \item{"ylim"}{  numeric vector of length 2, denoting the limits on the y axis of the plot. Only active for types "rating_curve","rating_curve_mean","f","beta","sigma_eps","residuals".}
}}
}
\description{
Print the results of a  object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
gplm0.fit <- gplm0(f,V316_river)
plot(gplm0.fit)
}
}
\seealso{
\code{\link{gplm0}} for fitting the gplm0 model,\code{\link{summary.gplm0}} for summaries, \code{\link{predict.gplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm0}} to help visualize the full posterior distributions.
}
