% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{plot.tournament}
\alias{plot.tournament}
\title{Plot comparison of models in tournament}
\usage{
\method{plot}{tournament}(x, type = "rating_curve")
}
\arguments{
\item{x}{an object of class "tournament"}

\item{type}{character denoting the type of plot desired. One of "tree","DIC","residuals","f","sigma_eps","rating_curve","rating_curve_log"}
}
\description{
Plot comparison of models in tournament
}
\seealso{
\code{\link{summary.torunament}} for summaries
}
