% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{plot_fun}
\alias{plot_fun}
\title{Plot bdrc model objects}
\usage{
plot_fun(x, ...)
}
\arguments{
\item{x}{an object of class "bplm0","bplm","bgplm0" or "bgplm"}

\item{...}{Arguments to be passed to other methods. The following arguments are supported:
\itemize{
 \item{\code{type}}{ a character denoting what type of plot should be drawn. Possible types are
                   \itemize{
                      \item{"rating_curve"}{ to plot the rating curve on original scale.}
                      \item{"rating_curve_mean"}{ to plot the rating curve on a log scale.}
                      \item{"f"}{ to plot the power-law exponent}
                      \item{"sigma_eps"}{ to plot the standard deviation on the data level}
                      \item{"residuals"}{ to plot the log residuals}
                      \item{"trace"}{ to plot trace plots of a list of one or more parameter. Requires parameter names as additional arguments, see below.}
                      \item{"histogram"}{ to plot histograms of a list of one or more parameter. Requires parameter names as additional arguments, see below.}
                   }}
\item{\code{transformed}}{ a logical value indicating whether the quantity should be plotted on a transformed scale used during the Bayesian inference. Defaults to FALSE.}
\item{\code{title}}{ title of the plot. Defaults to NULL, i.e. no title}
\item{}{ additional characters or character vectors denoting the parameters to plot. Used when type is "trace" or "histogram". Allowed values are the parameter names found in the summary of the model object. See Examples.}
}}

\item{type}{a character to denote the plot type. One of 'trace','histogram','rating_curve','rating_curve_mean','residuals'}
}
\value{
returns an object of class ggplot2 or Grob object
}
\description{
Visualize results from model ojbects in bdrc, bplm0, bplm, bgplm0,bgplm
}
