% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotmodel1.R
\name{plotmodel1}
\alias{plotmodel1}
\title{Plots the rating curve for model 1}
\usage{
plotmodel1(filename, eps = F, png = F, realscale = T, logscale = T,
  logresiduals = T, realresiduals = T)
}
\arguments{
\item{filename}{Input is a string with the name of the txt file containing the stage and flow data. See also input in \code{\link{clean}}.}

\item{eps}{TRUE or FALSE whether or not to save the plots as eps files or not.}

\item{png}{TRUE or FALSE whether or not to save the plots as png files or not.}

\item{realscale}{Logical constant, whether or not to plot the real scale image.}

\item{logscale}{Logical constant, whether or not to plot the log scale image.}

\item{logresiduals}{Logical constant, whether or not to plot the residuals for the log scale.}

\item{realresiduals}{Logical constant, whether or not to plot the residuals for the real scale.}
}
\value{
The output is a list which contains the ggplots the user has chosen to plot.
}
\description{
This function takes in the filename of the data that is used to calculate the ratingcurve It uses the \code{\link{clean}} function
and the \code{\link{model1BH}} function to calculate the rating curve and then plot it.
}

