% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{predict.bgplm0}
\alias{predict.bgplm0}
\title{Predict method for bgplm0 fit}
\usage{
\method{predict}{bgplm0}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class "bgplm0"}

\item{newdata}{a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.}

\item{...}{not used in this function}
}
\value{
numeric vector of discharge values for the stage values given in newdata
}
\description{
Print the results of a  object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bgplm0.fit <- bgplm0(f,V316_river,h_max=2)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(bgplm0.fit,newdata=seq(1,2,by=0.01))
}
}
\seealso{
\code{\link{bgplm0}} for fitting the bgplm0 model,\code{\link{summary.bgplm0}} for summaries, \code{\link{predict.bgplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.bgplm0}} to help visualize the full posterior distributions.
}
