% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{predict.bplm0}
\alias{predict.bplm0}
\title{Predict method for bplm0 fit}
\usage{
\method{predict}{bplm0}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class "bplm0"}

\item{newdata}{a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.}

\item{...}{not used in this function}
}
\value{
numeric vector of discharge values for the stage values given in newdata
}
\description{
Print the results of a  object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bplm0.fit <- bplm0(f,V316_river,h_max=2)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(bplm0.fit,newdata=seq(1,2,by=0.01))
}
}
\seealso{
\code{\link{bplm0}} for fitting the bplm0 model,\code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.bplm0}} to help visualize the full posterior distributions.
}
