% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{predict.plm0}
\alias{predict.plm0}
\alias{predict.plm}
\alias{predict.gplm0}
\alias{predict.gplm}
\title{Predict method for discharge rating curves}
\usage{
\method{predict}{plm0}(object, newdata = NULL, wide = FALSE, ...)

\method{predict}{plm}(object, newdata = NULL, wide = FALSE, ...)

\method{predict}{gplm0}(object, newdata = NULL, wide = FALSE, ...)

\method{predict}{gplm}(object, newdata = NULL, wide = FALSE, ...)
}
\arguments{
\item{object}{an object of class "plm0", "plm", "gplm0" or "gplm".}

\item{newdata}{a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.}

\item{wide}{a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.}

\item{...}{not used in this function}
}
\value{
an object of class "data.frame" with four columns, h (stage), lower (2.5\% posterior predictive quantile), median (50\% posterior predictive quantile), upper (97.5\% posterior predictive quantile). If wide=TRUE, a matrix as described above (see wide parameter) is returned.
}
\description{
Predict the discharge for given stage values based on a discharge rating curve model object.
}
\section{Functions}{
\itemize{
\item \code{predict.plm0}: Predict method for plm0

\item \code{predict.plm}: Predict method for plm

\item \code{predict.gplm0}: Predict method for gplm0

\item \code{predict.gplm}: Predict method for gplm
}}

\examples{
\dontrun{
data(krokfors)
set.seed(1)
plm0.fit <- plm0(formula=Q~W,data=krokfors,h_max=10)
#predict rating curve on a equally 10 cm spaced grid from 9 to 10 meters
predict(plm0.fit,newdata=seq(9,10,by=0.1))
#h     lower    median     upper
#1   9.0  2.440093  3.625679  5.341507
#2   9.1  2.969410  4.434736  6.596974
#3   9.2  3.626720  5.380490  8.025924
#4   9.3  4.284265  6.431505  9.560236
#5   9.4  5.069827  7.638437 11.406727
#6   9.5  5.936947  8.951891 13.309962
#7   9.6  6.848027 10.403692 15.458685
#8   9.7  7.949128 11.970401 17.813098
#9   9.8  9.057391 13.714896 20.606599
#10  9.9 10.232563 15.690309 23.470253
#11 10.0 11.546006 17.727467 26.802094
}
}
\seealso{
\code{\link{plm0}}, \code{\link{plm}}, \code{\link{gplm0}} and \code{\link{gplm}} for fitting a discharge rating curve and \code{\link{summary.plm0}}, \code{\link{summary.plm}}, \code{\link{summary.gplm0}} and \code{\link{summary.gplm}} for summaries. It is also useful to look at \code{\link{plot.plm0}}, \code{\link{plot.plm}}, \code{\link{plot.gplm0}} and \code{\link{plot.gplm}} to help visualize all aspects of the fitted discharge rating curve. Additionally, \code{\link{spread_draws}} and \code{\link{spread_draws}} help working directly with the MCMC samples.
}
