% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_u.R
\name{predict_u}
\alias{predict_u}
\title{Predictive values for unoberved stages}
\usage{
predict_u(param, RC)
}
\arguments{
\item{param}{A vector of samples of theta and samples of betas from MCMC. Theta is a vector containing c (stage at which discharge is zero), two hyperparameters sig_b^2 and phi_b
and six lambda parameters that affect the variance through the Bspline functions.}

\item{RC}{A list containing prior parameters, matrices and the data which are calculated in \code{\link{model2BH}}}
}
\value{
\itemize{
\item Vector containing predictive values ypo and values of beta for every stage measurement.
}
}
\description{
Calculates predictive values for unobserved stages
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}

