% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.bgplm}
\alias{print.bgplm}
\title{Print bgplm object}
\usage{
\method{print}{bgplm}(x, ...)
}
\arguments{
\item{x}{an object of class "bgplm"}
}
\description{
Print the results of a bgplm object
}
\examples{
data(V316_river)
f <- Q~W
bgplm.fit <- bgplm(f,V316_river)
print(bgplm.fit)
}
\seealso{
\code{\link{summary.bgplm}} for summaries, \code{\link{predict.bgplm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bgplm.plot}} to help visualize the full posterior distributions.
}
