% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.bgplm0}
\alias{print.bgplm0}
\title{Print bgplm0 object}
\usage{
\method{print}{bgplm0}(x, ...)
}
\arguments{
\item{x}{an object of class "bgplm0"}
}
\description{
Print the results of a bgplm0 object
}
\examples{
data(V316_river)
f <- Q~W
bgplm0.fit <- bgplm0(f,V316_river)
print(bgplm0.fit)
}
\seealso{
\code{\link{summary.bgplm0}} for summaries, \code{\link{predict.bgplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bgplm0.plot}} to help visualize the full posterior distributions.
}
