% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.bplm}
\alias{print.bplm}
\title{Print bplm object}
\usage{
\method{print}{bplm}(x, ...)
}
\arguments{
\item{x}{an object of class "bplm"}
}
\value{
bgplm0 returns an object of class "bplm"\cr\cr
}
\description{
Print the results of a bplm object
}
\examples{
data(V316_river)
f <- Q~W
bplm.fit <- bplm(f,V316_river)
print(bplm.fit)
}
\seealso{
\code{\link{summary.bplm}} for summaries, \code{\link{predict.bplm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm.plot}} to help visualize the full posterior distributions.
}
