% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.bplm0}
\alias{print.bplm0}
\title{Print bplm0 object}
\usage{
\method{print}{bplm0}(x, ...)
}
\arguments{
\item{x}{an object of class "bplm0"}
}
\description{
Print the results of a bplm0 object
}
\examples{
data(V316_river)
f <- Q~W
bplm.fit <- bplm(f,V316_river)
print(bplm0.fit)
}
\seealso{
\code{\link{bplm0}} for fitting the bplm0 model, \code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm0.plot}} to help visualize the full posterior distributions.
}
