% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_methods.R
\name{print.compare}
\alias{print.compare}
\title{Print Method for compare Objects}
\usage{
\method{print}{compare}(x, ...)
}
\arguments{
\item{x}{An object of class "compare".}

\item{...}{Additional arguments passed to print (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
This function prints a formatted table of model comparison results based on WAIC.
}
\seealso{
\code{\link{compare}} to run a model comparison, \code{\link{plot.compare}} for visualizing the model comparison.
}
