% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.gplm}
\alias{print.gplm}
\title{Print gplm object}
\usage{
\method{print}{gplm}(x, ...)
}
\arguments{
\item{x}{an object of class "gplm"}

\item{...}{not used in this function}
}
\description{
Print the results of a gplm object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
gplm.fit <- gplm(f,V316_river)
print(gplm.fit)
}
}
\seealso{
\code{\link{summary.gplm}} for summaries, \code{\link{predict.gplm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm}} to help visualize the full posterior distributions.
}
