% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.gplm0}
\alias{print.gplm0}
\title{Print gplm0 object}
\usage{
\method{print}{gplm0}(x, ...)
}
\arguments{
\item{x}{an object of class "gplm0"}

\item{...}{not used in this function}
}
\description{
Print the results of a gplm0 object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
gplm0.fit <- gplm0(f,V316_river)
print(gplm0.fit)
}
}
\seealso{
\code{\link{summary.gplm0}} for summaries, \code{\link{predict.gplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm0}} to help visualize the full posterior distributions.
}
