% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.plm}
\alias{print.plm}
\title{Print plm object}
\usage{
\method{print}{plm}(x, ...)
}
\arguments{
\item{x}{an object of class "plm"}

\item{...}{not used in this function}
}
\value{
gplm0 returns an object of class "plm"\cr\cr
}
\description{
Print the results of a plm object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
plm.fit <- plm(f,V316_river)
print(plm.fit)
}
}
\seealso{
\code{\link{summary.plm}} for summaries, \code{\link{predict.plm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm}} to help visualize the full posterior distributions.
}
