% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.plm0}
\alias{print.plm0}
\title{Print plm0 object}
\usage{
\method{print}{plm0}(x, ...)
}
\arguments{
\item{x}{an object of class "plm0"}

\item{...}{not used in this function}
}
\description{
Print the results of a plm0 object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
plm0.fit <- plm0(f,V316_river)
print(plm0.fit)
}
}
\seealso{
\code{\link{plm0}} for fitting the plm0 model, \code{\link{summary.plm0}} for summaries, \code{\link{predict.plm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm0}} to help visualize the full posterior distributions.
}
