% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spanga}
\alias{spanga}
\title{Spanga stream gauging station in Sweden}
\format{
A data frame with columns:
\describe{
 \item{W}{Measurements of water stage}
 \item{Q}{Measurements of water discharge}
}
}
\source{
Swedish Meteorological and Hydrological Institute \url{http://www.smhi.se}
}
\usage{
spanga
}
\description{
Paired stage (W) and discharge measurements (Q) from Spanga stream gauging station in Sweden
}
\keyword{datasets}
