% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_draws.R
\name{spread_draws}
\alias{spread_draws}
\title{Spread MCMC chain draws to data.frame on a wide format}
\usage{
spread_draws(mod, ..., transformed = F)
}
\arguments{
\item{mod}{an object of class "bplm0","bplm","bgplm0" or "bgplm".}

\item{...}{any number of character vectors containing valid names of parameters in the model or "rating_curve" and "rating_curve_mean". Also accepts "latent_parameters" and "hyperparameters".}

\item{transformed}{boolean value determining whether the output is to be represented on the transformed scale used for sampling in the MCMC chain or the original scale. Defaults to FALSE.}
}
\value{
Data frame with columns
\code{chain}
\code{iter}
\code{param}
\code{value}
}
\description{
Useful to convert MCMC chain draws of particular parameters or output from the model object to a wide format for further data wrangling
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bplm0.fit <- bplm0(f,V316_river)
hyp_samples <- spread_draws(bplm0.fit,'hyperparameters')
head(hyp_samples)
rating_curve_samples <- spread_draws(bplm0.fit,'rating_curve','rating_curve_mean')
head(rating_curve_samples)
}
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson, & Sigurdur M. Gardarsson. (2020). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling.
}
\seealso{
\code{\link{bplm0}},\code{\link{bplm}},\code{\link{bgplm0}},\code{\link{bgplm}} for further information on parameters
}
