% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.bgplm}
\alias{summary.bgplm}
\title{Summarizing bplm fit}
\usage{
\method{summary}{bgplm}(object, ...)
}
\arguments{
\item{object}{an object of class "bgplm"}

\item{...}{not used in this function}
}
\value{
bgplm0 returns an object of class "bplm"\cr\cr
}
\description{
Summarize the results of a bgplm object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bgplm.fit <- bgplm(f,V316_river)
summary(bgplm.fit)
}
}
\seealso{
\code{\link{bgplm}} for fitting the bgplm model,\code{\link{summary.bgplm}} for summaries, \code{\link{predict.bgplm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.bgplm}} to help visualize the full posterior distributions.
}
