% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.bgplm0}
\alias{summary.bgplm0}
\title{Summarizing bgplm0 fit}
\usage{
\method{summary}{bgplm0}(object, ...)
}
\arguments{
\item{object}{an object of class "bgplm0"}

\item{...}{not used in this function}
}
\value{
bgplm0 returns an object of class "bplm"\cr\cr
}
\description{
Summarize the results of a bgplm0 object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bgplm0.fit <- bgplm0(f,V316_river)
summary(bgplm0.fit)
}
}
\seealso{
\code{\link{bgplm0}} for fitting the bgplm0 model,\code{\link{summary.bgplm0}} for summaries, \code{\link{predict.bgplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.bgplm0}} to help visualize the full posterior distributions.
}
