% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.bplm}
\alias{summary.bplm}
\title{Summarizing bplm fit}
\usage{
\method{summary}{bplm}(object, ...)
}
\arguments{
\item{object}{an object of class "bplm"}

\item{...}{not used in this function}
}
\value{
bgplm0 returns an object of class "bplm"\cr\cr
}
\description{
Summarize the results of a bplm object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bplm.fit <- bplm(f,V316_river)
summary(bplm.fit)
}
}
\seealso{
\code{\link{bplm}} for fitting the bplm model,\code{\link{summary.bplm}} for summaries, \code{\link{predict.bplm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.bplm}} to help visualize the full posterior distributions.
}
