% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.gplm0}
\alias{summary.gplm0}
\title{Summarizing gplm0 fit}
\usage{
\method{summary}{gplm0}(object, ...)
}
\arguments{
\item{object}{an object of class "gplm0"}

\item{...}{not used in this function}
}
\value{
gplm0 returns an object of class "plm"\cr\cr
}
\description{
Summarize the results of a gplm0 object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
gplm0.fit <- gplm0(f,V316_river)
summary(gplm0.fit)
}
}
\seealso{
\code{\link{gplm0}} for fitting the gplm0 model,\code{\link{summary.gplm0}} for summaries, \code{\link{predict.gplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.gplm0}} to help visualize the full posterior distributions.
}
