% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.plm}
\alias{summary.plm}
\title{Summarizing plm fit}
\usage{
\method{summary}{plm}(object, ...)
}
\arguments{
\item{object}{an object of class "plm"}

\item{...}{not used in this function}
}
\value{
gplm0 returns an object of class "plm"\cr\cr
}
\description{
Summarize the results of a plm object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
plm.fit <- plm(f,V316_river)
summary(plm.fit)
}
}
\seealso{
\code{\link{plm}} for fitting the plm model,\code{\link{summary.plm}} for summaries, \code{\link{predict.plm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm}} to help visualize the full posterior distributions.
}
