% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.plm0}
\alias{summary.plm0}
\title{Summarizing plm0 fit}
\usage{
\method{summary}{plm0}(object, ...)
}
\arguments{
\item{object}{an object of class "plm0"}

\item{...}{Not used for this function}
}
\description{
Summarize the results of a plm0 object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
plm0.fit <- plm0(f,V316_river)
summary(plm0.fit)
}
}
\seealso{
\code{\link{plm0}} for fitting the plm0 model, \code{\link{predict.plm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm0}} to help visualize the full posterior distributions.
}
