% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{theme_bdrc}
\alias{theme_bdrc}
\title{Custom bdrc theme}
\usage{
theme_bdrc(scaling = 1, ...)
}
\arguments{
\item{scaling}{a numerical value which can be used to scale up or down the size of the text and titles of a plot that uses \code{theme_bdrc}. Defaults to 1.}

\item{...}{not used in this function}
}
\value{
returns a theme object for the package
}
\description{
Custom bdrc theme
}
\keyword{internal}
