% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V316_river}
\alias{V316_river}
\title{Data from x river in Iceland}
\format{
A data frame with columns:
\describe{
 \item{W}{Measurements of river stage}
 \item{Q}{Measurements of river discharge}
}
}
\source{
Icelandic Meteorological Office. \url{https://en.vedur.is/}
}
\usage{
V316_river
}
\description{
Data from x river in Iceland
}
\examples{
\dontrun{
 V316_river
}
}
\keyword{datasets}
