% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/W_unobserved.R
\name{W_unobserved}
\alias{W_unobserved}
\title{Unobserved stages}
\usage{
W_unobserved(W_unique, min = NA, max = NA)
}
\arguments{
\item{W_unique}{vector containing unique stages from river data.}

\item{min}{minimum stage of rating curve.}

\item{max}{maximum stage of rating curve.}
}
\value{
W_unobserved returns a list of vectors, W_u and W_u_tild. W_u is a vector of unobserved stage values
needed to make an equally spaced grid of stages. W_u_tild is a vector which is calculated by W_u-min(W_unique) needed to input into B_splines.
The unobserved stages are lower or higher than that of the data, take the same value in W_u_tild as the minimum value and maximum value of the
data respectively. This is done to ensure constant variance below and above observed data.
}
\description{
W_unobserved returns the stages that are needed to make an equally spaced grid of stages from data of stages.
}

