% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Adist.R
\name{Adist}
\alias{Adist}
\title{Linking unique water level measurements to actual
water level measurements}
\usage{
Adist(w)
}
\arguments{
\item{w}{Numerical vector containing water level measurements}
}
\value{
\itemize{
\item A: Matrix \strong{A} linking unique water level measurements (\strong{w'}) to actual
water level measurements (w) such that \strong{w}=\strong{Aw'}
\item dist:  Matrix of distances between unique water level measurements
\item n:     Number of unique measurements
\item N:     Number of measurements
}
}
\description{
Adist links unique water level measurements (\strong{w'}) to actual
water level measurements (w) such that \strong{w}=\strong{Aw'}.
from the measurements.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}

