% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{autoplot.plm0}
\alias{autoplot.plm0}
\alias{autoplot.plm}
\alias{autoplot.gplm0}
\alias{autoplot.gplm}
\title{Autoplot method for discharge rating curves}
\usage{
\method{autoplot}{plm0}(
  object,
  ...,
  type = "rating_curve",
  param = NULL,
  transformed = FALSE,
  title = NULL,
  xlim = NULL,
  ylim = NULL
)

\method{autoplot}{plm}(
  object,
  ...,
  type = "rating_curve",
  param = NULL,
  transformed = FALSE,
  title = NULL,
  xlim = NULL,
  ylim = NULL
)

\method{autoplot}{gplm0}(
  object,
  ...,
  type = "rating_curve",
  param = NULL,
  transformed = FALSE,
  title = NULL,
  xlim = NULL,
  ylim = NULL
)

\method{autoplot}{gplm}(
  object,
  ...,
  type = "rating_curve",
  param = NULL,
  transformed = FALSE,
  title = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{object}{an object of class "plm0","plm","gplm0" or "gplm".}

\item{...}{other plotting parameters (not used in this function)}

\item{type}{a character denoting what type of plot should be drawn. Defaults to "rating_curve". Possible types are
\itemize{
   \item{"rating_curve"}{ to plot the rating curve.}
   \item{"rating_curve_mean"}{ to plot the posterior mean of the rating curve.}
   \item{"f"}{ to plot the power-law exponent.}
   \item{"beta"}{ to plot the random effect in the power-law exponent.}
   \item{"sigma_eps"}{ to plot the standard deviation on the data level.}
   \item{"residuals"}{ to plot the log residuals.}
   \item{"trace"}{ to plot trace plots of parameters given in param.}
   \item{"histogram"}{ to plot histograms of parameters given in param.}
}}

\item{param}{a character vector with the parameters to plot. Defaults to NULL and is only used if type is "trace" or "histogram". Allowed values are the parameters given in the model summary of x as well as "hyperparameters" or "latent_parameters" for specific groups of parameters.}

\item{transformed}{a logical value indicating whether the quantity should be plotted on a transformed scale used during the Bayesian inference. Defaults to FALSE.}

\item{title}{a character denoting the title of the plot}

\item{xlim}{numeric vector of length 2, denoting the limits on the x axis of the plot. Applicable for types "rating_curve","rating_curve_mean","f","beta","sigma_eps","residuals".}

\item{ylim}{numeric vector of length 2, denoting the limits on the y axis of the plot. Applicable for types "rating_curve","rating_curve_mean","f","beta","sigma_eps","residuals".}
}
\value{
returns an object of class "ggplot2".
}
\description{
Visualize discharge rating curve model objects
}
\section{Functions}{
\itemize{
\item \code{autoplot(plm0)}: Autoplot method for plm0

\item \code{autoplot(plm)}: Autoplot method for plm

\item \code{autoplot(gplm0)}: Autoplot method for gplm0

\item \code{autoplot(gplm)}: Autoplot method for gplm

}}
\examples{
\donttest{
library(ggplot2)
data(krokfors)
set.seed(1)
plm0.fit <- plm0(Q~W,krokfors,num_cores=2)
autoplot(plm0.fit)
autoplot(plm0.fit,transformed=TRUE)
autoplot(plm0.fit,type='histogram',param='c')
autoplot(plm0.fit,type='histogram',param='c',transformed=TRUE)
autoplot(plm0.fit,type='histogram',param='hyperparameters')
autoplot(plm0.fit,type='histogram',param='latent_parameters')
autoplot(plm0.fit,type='residuals')
autoplot(plm0.fit,type='f')
autoplot(plm0.fit,type='sigma_eps')
}
}
\seealso{
\code{\link{plm0}}, \code{\link{plm}}, \code{\link{gplm0}} and \code{\link{gplm}} for fitting a discharge rating curve and \code{\link{summary.plm0}}, \code{\link{summary.plm}}, \code{\link{summary.gplm0}} and \code{\link{summary.gplm}} for summaries. It is also useful to look at \code{\link{spread_draws}} and \code{\link{gather_draws}} to work directly with the MCMC samples.
}
