% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{autoplot.tournament}
\alias{autoplot.tournament}
\title{Autoplot method for discharge rating curve tournament}
\usage{
\method{autoplot}{tournament}(x, type = "deviance", ...)
}
\arguments{
\item{x}{an object of class "tournament"}

\item{type}{a character denoting what type of plot should be drawn. Possible types are
\itemize{
 \item{"deviance"}{ to plot the deviance of the four models.}
}}

\item{...}{further arguments passed to other methods.}
}
\description{
Compare the four discharge rating curves from the tournament object in different ways
}
\examples{
\donttest{
data(krokfors)
t_obj <- tournament(Q~W,krokfors)
autoplot(t_obj)
}
}
\seealso{
\code{\link{tournament}} to run a dishcarge rating curve tournament and \code{\link{summary.tournament}} for summaries.
}
