% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{autoplot.tournament}
\alias{autoplot.tournament}
\title{Autoplot - Comparison of models in tournament}
\usage{
\method{autoplot}{tournament}(x, type = "deviance", ...)
}
\arguments{
\item{x}{an object of class "tournament"}

\item{type}{a character denoting what type of plot should be drawn. Possible types are
\itemize{
 \item{"deviance"}{ to plot the rating curve on original scale.}
}}

\item{...}{not used in this function}
}
\description{
Compare the four models from the tournament object in different ways
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
t_obj <- tournament(f,V316_river)
autoplot(t_obj)
}
}
\seealso{
\code{\link{summary.tournament}} for summaries
}
