% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgplm0.R
\name{bgplm0}
\alias{bgplm0}
\title{Bayesian Generalized Power Law Model with Constant Variance}
\usage{
bgplm0(
  formula,
  data,
  c_param = NULL,
  w_max = NULL,
  forcepoint = rep(FALSE, nrow(data)),
  ...
)
}
\arguments{
\item{formula}{formula with name of discharge column in data as response and name of stage column in data as the single covariate.}

\item{data}{data.frame containing the columns in formula}

\item{w_max}{vector of length 2 setting the lower and upper bound of stage values at which a rating curve should be predicted. If NULL, the known value of c or the mle of c will be used as lower bound (depending on the value of the input parameter c) and maximum stage value in data as upper bound.}

\item{country}{Name of the country the prior parameters should be defined for, default value is "Iceland".}

\item{Wmin}{Positive numeric value for the lowest stage the user wants to calculate a rating curve. If input is an empty string (default) Wmin will
automatically be set to c_hat.}

\item{Wmax}{Positive numeric value for the highest stage the user wants to calculate a rating curve. If input is an empty string (default) Wmax will
automatically be set to the maximum stage of the data.}
}
\value{
List containing information on the calculated rating curve,
the data frames observedData, betaData, completePrediction, observedPrediction, TableOfData, FitTable, LowerTable, UpperTable, plotTable.
}
\description{
Infers a rating curve for paired measurements of stage and discharge using a generalized power law model described in Hrafnkelsson et al.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
\seealso{
\code{\link{clean}}
}
