% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplm0.R
\name{bplm0}
\alias{bplm0}
\title{Bayesian Power Law Model with Constant Variance}
\usage{
bplm0(
  formula,
  data,
  c_param = NULL,
  w_max = NULL,
  forcepoint = rep(FALSE, nrow(data)),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula", with discharge column name as response and stage column name as a covariate.The details of model specification are given under "Details".}

\item{data}{data.frame containing the variables specified in formula}

\item{c_param}{stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data}

\item{w_max}{maximum stage to which the rating curve should extrapolate for. If NULL, the maximum stage value in data is selected as an upper bound.}

\item{forcepoint}{A boolean vector of the same length as the number of rows in data. If an element at index i is TRUE it indicates that the rating curve should be forced through the i-th measurement. Use with care, as this will strongly influence the resulting rating curve.}
}
\value{
bplm0 returns an object of class "bplm0"
The functions summary are used to obtain an print summary of the fitted model.
An object of class "bplm0" is a list containt at least the following component:
\code{rating_curve} \cr
\code{rating_curve_mean} \cr
\code{param_summary} \cr
\code{DIC_summary} \cr
\code{Q_posterior_predictive} \cr
\code{Q_posterior} \cr
\code{a_posterior} \cr
\code{b_posterior} \cr
\code{c_posterior} \cr
\code{sigma_eps_posterior} \cr
\code{formula} \cr
\code{data} \cr
\code{run_info}
}
\description{
bplm0 is used to fit a rating curve for paired measurements of stage and discharge using a Bayesian power law model with constant variance as described in Hrafnkelsson et al.
}
\examples{
data(sim_dat)
f <- Q~W
bplm0.fit <- bplm0(f,sim_dat)
summary(bplm9.fit)
plot(bplm0.fit)
bplm0.fit_known_c <- bplm0(f,sim_dat,c_param=)
summary(bplm9.fit)
plot(bplm0.fit)
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson, & Sigurdur M. Gardarsson. (2020). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling.
}
\seealso{
\code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm0.plot}} to help visualize the quality of inference.
}
