% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Cleans the input file}
\usage{
clean(file, advanced = FALSE, slider = 0, dummy = NULL, keeprows = NULL,
  force = NULL, shiny = FALSE, Wmin = NA, Wmax = NA, experiod = FALSE,
  dates = 0)
}
\arguments{
\item{file}{Is a string that contains the name of a txt file that contains stage and flow data from a certain river}

\item{advanced}{TRUE or FALSE depending if you want to use the advanced settings}

\item{slider}{Input should be a vector with two integers. The integers span the date constrainment for the data.
So every datapoint that is not inside that date range will be discarded.}

\item{dummy}{}

\item{keeprows}{}

\item{force}{}

\item{shiny}{}

\item{Wmin}{}

\item{Wmax}{}

\item{experiod}{TRUE or FALSE depending on whether you want to exclude a date range from the data}

\item{dates}{Input is a vector with two Date values. The dates span the range that you want to exclude from the data}
}
\value{
If all the parameters are either default or used as described the output will be a list containing a matrix with the cleaned stage
and flow values as well as a data frame with the cleaned data cointaining date, time, quality, stage and flow.
}
\description{
This function takes in the data and cleans it, if advanced is TRUE and experiod is TRUE it conditions and fixes
the data depending on the input of the user.
}

