% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{create_A}
\alias{create_A}
\title{Linking unique water level measurements to actual
water level measurements}
\usage{
create_A(w)
}
\arguments{
\item{w}{numeric vector of stage measurements in meters}
}
\value{
\itemize{
\item A: Matrix \strong{A} linking unique water level measurements (\strong{w'}) to actual
water level measurements (w) such that \strong{w}=\strong{Aw'}
}
}
\description{
Adist links unique water level measurements (\strong{w'}) to actual
water level measurements (w) such that \strong{w}=\strong{Aw'}.
from the measurements.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
