% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report_pages.plm}
\alias{get_report_pages.plm}
\title{Get plm fit report pages}
\usage{
\method{get_report_pages}{plm}(x, type = 1, ...)
}
\arguments{
\item{x}{an object of class "plm".}

\item{type}{an integer to determine whether to get the pages from report type 1 or 2. Only type 1 is permissible for an object of class "plm" which produces a list of two grob objects, the first grob object consists of a plot collage and parameter summary table, the second a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Get a list of the grob objects used to by \code{\link{get_report}} create a pdf report.
}
\examples{
\dontrun{
data(lisjobacken)
f <- Q~W
plm.fit <- plm(f,lisjobacken)
grob_object_list <- get_report_pages(plm.fit)
}
}
\seealso{
\code{\link{plm}} for fitting the plm model,\code{\link{summary.plm}} for summaries, \code{\link{predict.plm}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm}} to help visualize the full posterior distributions.
}
