% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_report_pages.plm0}
\alias{get_report_pages.plm0}
\title{Get plm0 fit report pages}
\usage{
\method{get_report_pages}{plm0}(x, type = 1, ...)
}
\arguments{
\item{x}{an object of class "plm0".}

\item{type}{an integer to determine whether to get the pages from report type 1 or 2. Only type 1 is permissible for an object of class "plm0" which produces a list of two grob objects, the first grob object consists of a plot collage and parameter summary table, the second a tabular rating curve.}

\item{...}{further arguments passed to other methods (currently unused).}
}
\description{
Get a list of the grob objects used by \code{\link{get_report}} to create a pdf report.
}
\examples{
\dontrun{
data(halla)
f <- Q~W
plm0.fit <- plm0(f,halla)
grob_object_list <- get_report_pages(plm0.fit)
}
}
\seealso{
\code{\link{plm0}} for fitting the plm0 model,\code{\link{summary.plm0}} for summaries, \code{\link{predict.plm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm0}} to help visualize the full posterior distributions.
}
