% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model1.R
\name{model1BH}
\alias{model1BH}
\title{Calculates the rating curve using model1}
\usage{
model1BH(clean, country = "Iceland", Wmin = "", Wmax = "")
}
\arguments{
\item{clean}{Is the output from the \code{\link{clean}} function. That is a list containg a matrix and a data frame.
The matrix containing the stage(in meters) and flow values and the data frame containing date, time, quality, stage and flow values.}

\item{country}{The input is a string with the name of the country you want to use the prior parameters from, e.g. country='Iceland'.}

\item{Wmin}{The input is either the default empty string or an integer for the lowest stage level. If you set as.numeric("") you get Na and then Wmin will be
automatically set to c_hat.}

\item{Wmax}{The input is either the default empty string or an integer for the highest stage level. If you set as.numeric("") you get Na and then Wmax will be
automatically set to the maximum stage of the data.}
}
\value{
If all the parameters are used as described the output will be a list containing the integers varappr and c_hat, the matrix mu and
the data frames qvdata, simdata, realdata , tafla, fitrctafla, lowerrctafla, upperrctafla, plottafla.
}
\description{
This function takes in the cleaned data from the \code{\link{clean}} function
and calculates the rating curve using model 1.
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
\seealso{
\code{\link{clean}}
}

