% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm0.R
\name{plm0}
\alias{plm0}
\title{Power Law Model with constant variance}
\usage{
plm0(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  parallel = T,
  forcepoint = rep(FALSE, nrow(data))
)
}
\arguments{
\item{formula}{an object of class "formula", with discharge column name as response and stage column name as a covariate, i.e. of the form y~x where y is discharge in \eqn{m^3/s} and x is stage in \eqn{m} (it is very important that the data is in the correct units).}

\item{data}{data.frame containing the variables specified in formula.}

\item{c_param}{stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data.}

\item{h_max}{maximum stage to which the rating curve should extrapolate to. If NULL, the maximum stage value in data is selected as an upper bound.}

\item{parallel}{logical value indicating whether to run the MCMC in parallel or not. Defaults to TRUE.}

\item{forcepoint}{logical vector of the same length as the number of rows in data. If an element at index i is TRUE it indicates that the rating curve should be forced through the i-th measurement. Use with care, as this will strongly influence the resulting rating curve.}
}
\value{
plm0 returns an object of class "plm0". An object of class "plm0" is a list containing the following components: \cr
\item{\code{rating_curve}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of the rating curve.}
\item{\code{rating_curve_mean}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of the mean of the rating curve.}
\item{\code{param_summary}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of latent- and hyperparameters.}
\item{\code{Deviance_summary}}{a data frame with 2.5\%, 50\% and 97.5\% quantiles of the posterior distribution of the deviance.}
\item{\code{rating_curve_posterior}}{a matrix containing the full thinned posterior samples of the posterior distribution of the rating curve (excluding burn-in).}
\item{\code{rating_curve_mean_posterior}}{a matrix containing the full thinned posterior samples of the posterior distribution of the mean of the rating curve (excluding burn-in).}
\item{\code{a_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{a}.}
\item{\code{b_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{b}.}
\item{\code{c_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{c}.}
\item{\code{sigma_eps_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of \eqn{\sigma_{\epsilon}}.}
\item{\code{Deviance_posterior}}{a numeric vector containing the full thinned posterior samples of the posterior distribution of the deviance excluding burnin samples.}
\item{\code{D_hat}}{deviance at the median value of the parameters}
\item{\code{num_effective_param}}{number of effective parameters, which is calculated as median(Deviance_posterior) - D_hat}
\item{\code{DIC}}{Deviance Information Criterion for the model, calculated as D_hat + 2*num_effective_parameters}
\item{\code{acceptance_rate}}{proportion of accepted samples in the thinned MCMC chain (excluding burn-in).}
\item{\code{formula}}{object of type "formula" provided by the user.}
\item{\code{data}}{data provided by the user.}
\item{\code{run_info}}{Information about the specific parameters used in the MCMC chain.}
}
\description{
plm0 is used to fit a discharge rating curve for paired measurements of stage and discharge using a Power Law Model with constant variance as described in Hrafnkelsson et al. See "Details" for a more elaborate description of the model.
}
\details{
The power-law model, which is commonly used in hydraulic practice, is of the form
\deqn{Q=a(h-c)^{b}}
where \eqn{Q} is discharge, \eqn{h} is stage and \eqn{a}, \eqn{b} and \eqn{c} are unknown constants. It is presented here as a Bayesian hierarchical model. The model is on a logarithmic scale
\deqn{log(Q_i) = log(a) + b log(h_i - c) + \epsilon,     i = 1,...,n}
where \eqn{\epsilon} follows a normal distribution with mean zero and variance \eqn{\sigma_\epsilon^2}, independent of stage. An efficient posterior simulation is achieved by sampling from the joint posterior density of the hyperparameters of the model, and then sampling from the density of the latent parameters conditional on the hyperparameters.\cr\cr
Bayesian inference is based on the posterior density and summary statistics such as the posterior mean and 95\% posterior intervals are based on the posterior density. Analytical formulas for these summary statistics are intractable in most cases and thus they are computed by generating samples from the posterior density using a Markov chain Monte Carlo simulation.
}
\examples{
\dontrun{
data(V316_river)
set.seed(1)
formula <- Q~W
plm0.fit <- plm0(formula,V316_river)
summary(plm0.fit)
plot(plm0.fit)
plm0.fit_known_c <- plm0(formula,V316_river,c_param=0.75,h_max=2)
summary(plm0.fit_known_c)
plot(plm0.fit_known_c)
}
}
\references{
B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv
preprint 2010.04769
}
\seealso{
\code{\link{summary.plm0}} for summaries, \code{\link{predict.plm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.plm0}} to help visualize the full posterior distributions.
}
