% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotmodel2.R
\name{plotmodel2}
\alias{plotmodel2}
\title{Graphical rating curve for model 1}
\usage{
plotmodel2(filename, eps = F, png = F, realscale = T, logscale = T,
  standardResiduals = T, realresiduals = T, beta = T)
}
\arguments{
\item{filename}{Input is a string with the name of the txt file containing data. See also input in \code{\link{clean}}.}

\item{eps}{TRUE or FALSE whether or not to save the plots as eps files or not.}

\item{png}{TRUE or FALSE whether or not to save the plots as png files or not.}

\item{realscale}{Logical, whether or not to plot the real scale rating curve.}

\item{logscale}{Logical, whether or not to plot the log scale rating curve.}

\item{standardResiduals}{Logical, whether or not to plot the standardized residuals.}

\item{realresiduals}{Logical, whether or not to plot the real scale residuals.}
}
\value{
Returns a list of ggplot objects that the user chose to plot.
}
\description{
Takes in name of a stage-discharge file, calculates a rating curve using \code{\link{model2BH}} and returns ggplot figures of custom outputs
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
\seealso{
\code{\link{clean}} , \code{\link{model2BH}}
}

