% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{predict.bgplm}
\alias{predict.bgplm}
\title{Predict method for bplm0 fit}
\usage{
\method{predict}{bgplm}(object, newdata = NULL)
}
\arguments{
\item{object}{an object of class "bplm0"}

\item{newdata}{a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.}
}
\value{
numeric vector of discharge values for the stage values given in newdata
}
\description{
Print the results of a  object
}
\examples{
data(V316_river)
f <- Q~W
bplm0.fit <- bplm0(f,V316_river)
plot(bplm0.fit)
}
\seealso{
\code{\link{bplm0}} for fitting the bplm0 model,\code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm0.plot}} to help visualize the full posterior distributions.
}
