% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{print.bplm0}
\alias{print.bplm0}
\title{Print bplm0 object}
\usage{
\method{print}{bplm0}(x, ...)
}
\arguments{
\item{x}{an object of class "bplm0"}

\item{...}{not used in this function}
}
\description{
Print the results of a bplm0 object
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
bplm0.fit <- bplm0(f,V316_river)
print(bplm0.fit)
}
}
\seealso{
\code{\link{bplm0}} for fitting the bplm0 model, \code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{plot.bplm0}} to help visualize the full posterior distributions.
}
