% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{print.tournament}
\alias{print.tournament}
\title{Print tournament object}
\usage{
\method{print}{tournament}(x)
}
\arguments{
\item{x}{an object of class "tournament"}
}
\description{
Print the results of a tournament of model comparisons
}
\seealso{
\code{\link{summary.torunament}} for summaries
}
