% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{print.tournament}
\alias{print.tournament}
\title{Print method for discharge rating curve tournament}
\usage{
\method{print}{tournament}(x, ...)
}
\arguments{
\item{x}{An object of class "tournament"}

\item{...}{Not used in this function}
}
\description{
Print the results of a tournament of discharge rating curve model comparisons
}
\seealso{
\code{\link{tournament}} to run a discharge rating curve tournament, \code{\link{summary.tournament}} for summaries and \code{\link{plot.tournament}} for visualizing the mode comparison.
}
