% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\title{Chooses priors for the model}
\usage{
priors(country)
}
\arguments{
\item{country}{Is a string with the name of the country the data is from.}
}
\value{
The output is the RC list which contains the priors and the outcomes of the prior calculations.
}
\description{
Takes in the country as a string and chooses which priors to use and then makes the prior calculations
}

