% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{priors}
\alias{priors}
\title{Prior parameter specification}
\usage{
priors(model, c_param)
}
\arguments{
\item{country}{A string with the name of the country of which the prior parameters into the models should be specified for}
}
\value{
The priors are based from data from rivers of a given country.If you want to add your country to this function,
}
\description{
Prior parameter specification
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
