% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{priors}
\alias{priors}
\title{Prior parameter specification}
\usage{
priors(model, c_param = NULL)
}
\arguments{
\item{model}{character denoting which model to get the fixed values of parameters of prior distributions. One of plm0,plm,gplm0,gplm}

\item{c_param}{The value of the parameter c, the stage at zero discharge. NULL if unknown}
}
\value{
The priors are based from data from rivers of a given country.If you want to add your country to this function,
}
\description{
Prior parameter specification
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson and Sigurdur M. Gardarson (2015) \emph{Bayesian Generalized Rating Curves}
}
