% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skogsliden}
\alias{skogsliden}
\title{Skogsliden gauging station in Sweden}
\format{
A data frame with columns:
\describe{
 \item{\code{W}}{Measurements of water stage in meters}
 \item{\code{Q}}{Measurements of water discharge in cubic meters per second}
}
}
\source{
Swedish Meteorological and Hydrological Institute.
}
\usage{
skogsliden
}
\description{
Data on discharge and stage from Skogsliden gauging station in Sweden
}
\keyword{datasets}
