% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spanga}
\alias{spanga}
\title{Spanga stream gauging station in Sweden}
\format{
A data frame with columns:
\describe{
 \item{W}{Measurements of water stage in meters}
 \item{Q}{Measurements of water discharge in cubic meters per second}
}
}
\source{
Swedish Meteorological and Hydrological Institute.
}
\usage{
spanga
}
\description{
Data on discharge and stage from Spanga stream gauging station in Sweden.
}
\keyword{datasets}
