% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_draws.R
\name{spread_draws}
\alias{spread_draws}
\title{Spread MCMC chain draws to data.frame}
\usage{
spread_draws(mod, param, transformed = F)
}
\arguments{
\item{mod}{an object of class "bplm0","bplm","bgplm0" or "bgplm".}

\item{param}{character with the name of the parameter of interest. Also accepts "latent_parameters" and "hyperparameters".}

\item{transformed}{boolean value determining whether the parameter is to be represented on the transformed scale used in sampling or the original scale. Defaults to FALSE.}
}
\value{
Data frame with columns
\code{chain}
\code{iter}
\code{param}
\code{value}
}
\description{
Useful to convert MCMC chains draws of particular parameters in a model object to a tidyverse friendly data frame for further data wrangling and plotting
}
\examples{
data(sim_dat)
f <- Q~W
bplm0.fit <- bplm0(f,sim_dat)
summary(bplm9.fit)
plot(bplm0.fit,type='rating_curve')
}
\references{
Birgir Hrafnkelsson, Helgi Sigurdarson, & Sigurdur M. Gardarsson. (2020). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling.
}
\seealso{
\code{\link{bplm0}},\code{\link{bplm}},\code{\link{bgplm0}},\code{\link{bgplm}} for further information on parameters
}
