% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_methods.R
\name{summary.bplm}
\alias{summary.bplm}
\title{Summarizing bplm fit}
\usage{
\method{summary}{bplm}(object, ...)
}
\arguments{
\item{x}{an object of class "bplm0"}
}
\value{
bgplm0 returns an object of class "bplm"\cr\cr
}
\description{
Print the results of a  object
}
\examples{
data(V316_river)
f <- Q~W
bplm0.fit <- bplm0(f,V316_river)
summary(bplm0.fit)
}
\seealso{
\code{\link{bplm0}} for fitting the bplm0 model,\code{\link{summary.bplm0}} for summaries, \code{\link{predict.bplm0}} for prediction. It is also useful to look at \code{\link{spread_draws}} and \code{\link{bplm0.plot}} to help visualize the full posterior distributions.
}
