% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{summary.tournament}
\alias{summary.tournament}
\title{Print summary of tournament object}
\usage{
\method{summary}{tournament}(object, ...)
}
\arguments{
\item{object}{an object of class "tournament"}

\item{...}{not used in this function}
}
\description{
Print the summary of a tournament of model comparisons
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
t_obj <- tournament(f,V316_river)
summary(t_obj)
}
}
