% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{summary.tournament}
\alias{summary.tournament}
\title{Print summary of tournament object}
\usage{
\method{summary}{tournament}(x)
}
\arguments{
\item{x}{an object of class "tournament"}
}
\description{
Print the summary of a tournament of model comparisons
}
