% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_methods.R
\name{summary.tournament}
\alias{summary.tournament}
\title{Summary method for a discharge rating curve tournament}
\usage{
\method{summary}{tournament}(object, ...)
}
\arguments{
\item{object}{an object of class "tournament"}

\item{...}{not used in this function}
}
\description{
Print the summary of a tournament of model comparisons
}
\examples{
\donttest{
data(krokfors)
set.seed(1)
t_obj <- tournament(Q~W,krokfors,num_cores=2)
summary(t_obj)
}
}
\seealso{
\code{\link{tournament}} to run a discharge rating curve tournament and \code{\link{plot.tournament}} for visualizing the mode comparison
}
