% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{tournament}
\alias{tournament}
\title{Determine the most adequate rating curve model}
\usage{
tournament(formula = NULL, data = NULL, c_param = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula", with discharge column name as response and stage column name as a covariate.}

\item{data}{data.frame containing the variables specified in formula.}

\item{...}{if data and formula are set to NULL, one can add four model objects of types "gplm", "gplm0", "plm" and "plm0". This prevents the function from running all four models explicitly.}
}
\value{
An object of type "tournament" with the following elements
\describe{
 \item{\code{summary}}{a data frame with information on reults of the different games in the tournament.}
 \item{\code{contestants}}{model objects of types "plm0","plm","bpglm0" and "gplm" being compared.}
 \item{\code{winner}}{model object of the tournament winner.}
}
}
\description{
tournament compares four rating curve models of different complexities and determines which model is the most adequate
}
\details{
TODO
}
\examples{
\dontrun{
data(V316_river)
f <- Q~W
t_obj <- tournament(f,V316_river)
plot(t_obj)
}
}
\references{
B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv preprint 2010.04769.
}
\seealso{
\code{\link{summary.tournament}}
}
