% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{tournament}
\alias{tournament}
\title{Determine the most adequate rating curve model}
\usage{
tournament(formula = NULL, data = NULL, ..., winning_criteria = 0.75)
}
\arguments{
\item{formula}{an object of class "formula", with discharge column name as response and stage column name as a covariate.}

\item{data}{data.frame containing the variables specified in formula.}

\item{...}{if data and formula are set to NULL, one can add four model objects of types "gplm", "gplm0", "plm" and "plm0". This prevents the function from running all four models explicitly.}

\item{winning_criteria}{a numerical value between 0 and 1 which sets the threshold for which the probability of the more complex model in each model comparison, given the data and calculated using Bayes factor, must exceed for it to be declared the better model. This value defaults to 0.75 to favor the less complex models when the superiority of the more complex model is somewhat ambiguous.}
}
\value{
An object of type "tournament" with the following elements
\describe{
 \item{\code{summary}}{a data frame with information on reults of the different games in the tournament.}
 \item{\code{contestants}}{model objects of types "plm0","plm","gplm0" and "gplm" being compared.}
 \item{\code{winner}}{model object of the tournament winner.}
}
}
\description{
tournament compares four rating curve models of different complexities and determines which model is the most adequate
}
\details{
Tournament is a comparison method based on three matches that uses Bayes factor to compute the posterior probabilities of the models and select the most appropriate of the four models given the data. The first round of model comparisons sets up model types "gplm" vs. "gplm0" and "plm" vs. "plm0". If the posterior probability of the more complex model ("gplm" and "plm", respectively) exceeds the "winning_criteria" (defult value = 0.75) then it is chosen as the more appropriate model and moves on to the second and final round, where the winners from the first round will be compared in the same way. In the second round, if the more complex model (now the generalized power-law model) exceeds the same "winning_criteria" then it is chosen as the overall tournament winner and deemed the most appropriate model given the data. In each of the three matches, the posterior probabilities of the models are computed using the Bayes factor and assuming a priori that the two models were equally likely [see Jeffreys (1961) and Kass and Raftery (1995)].
}
\examples{
\donttest{
data(krokfors)
t_obj <- tournament(Q~W,krokfors)
plm0.fit <- plm0(Q~W,krokfors)
plm.fit <- plm(Q~W,krokfors)
gplm0.fit <- gplm0(Q~W,krokfors)
gplm.fit <- gplm(Q~W,krokfors)
t_obj <- tournament(gplm.fit,gplm0.fit,plm.fit,plm0.fit,winning_criteria=0.95)
}
}
\references{
B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv preprint 2010.04769.

Jeffreys, H. (1961). Theory of Probability, Third Edition. Oxford University Press.

Kass, R., and A. Raftery, A. (1995) Bayes Factors. Journal of the American Statistical Association, 90, 773-795.
}
\seealso{
\code{\link{summary.tournament}} and \code{\link{plot.tournament}}
}
