% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V316_river}
\alias{V316_river}
\title{Data from x river in Iceland}
\format{
A data frame with columns:
\describe{
 \item{W}{Measurements of river stage}
 \item{Q}{Measurements of river discharge}
}
}
\source{
Icelandic Meteorological Office. \url{http://www.vedur.is}
}
\usage{
V316_river
}
\description{
Paired stage and discharge measurements from x to y
}
\keyword{datasets}
